Add abstract method Node#fillFormFields(String, Map) to facilitate form-filling, change other classes accordingly.
The idea is to provide a serialized object containing the user's selection and a map that maps field-names to data. The
fillFormFields method deserializes the object to get access to the input data, and then matches up the input with the
proper form-field names. Once all fields are filled, the map can be used to apply the data to an Acrobat form.

Replace "option" with "choice" in body of ChoiceNode to increase language consistency.

Add fieldName to ChoiceNode, and choiceName to each Choice, for identifying the form-fields to fill.

Implement fillFormFields in FieldsNode and NoteNode.

Adjust ChangeOrderForm according to changes listed above.