Store decisions and output form data using a queue
The idea with the queue is to track where the user goes in the tree.
Once they finish making a path through the tree, I can simply empty the
queue in order and output the data for each decision they made.

A problem arises if the user goes back to change anything. The queue
will just keep growing, rather than replacing the path they took with a
different one. This will be addressed in another commit that will
likely result in the removal of the queue altogether. I just wanted to
preserve the idea here as a stepping stone in the path toward my final
solution.

DecisionStore is an API for storing DecisionQueues. The idea is that
many users may be using the system, so we want to have a DecisionQueue
for each of them. It's generic, so it can be backed by a Map, a
database, or whatever.