Implement "decision tree" to track decisions and retrieve form output
Did away with DecisionQueue because it didn't handle the case in which
the user goes back and changes their mind on something. I did a few
things to fix this:

1. Create a DecisionMap, in which each node is mapped to a
corresponding Decision as it is visited. This way, I never lose any of
the decisions made by the user. If they revisit a decision, that may
result in them taking a whole new branch down the decision tree.
Storing them in a map means that they can come back and change their
mind back to what they had originally, and all the decisions they made
will still be there.

2. Add previous and next pointers to each Decision. The Decision
already has a context, which denotes the node that the decision applies
to. The previous and next pointers allow me to traverse the most
recently active path through the "decision tree" as long as they are
updated each time the user makes a decision.

Together these changes allow me to record all the decisions made by the
user, traverse their path through the decision tree, and come up with
the correct form output based on that path, excluding any extra stuff
they may have entered by visiting a different path in the tree.