Implement CalculationNode, fix bugs
The implementation of CalculationNodes required me to make changes to
all of the Nodes as well as how decisions are made.

Added FilledFormFields to the data we're storing, so that
CalculationNodes can reference previously filled form fields.

Had to rewrite the storage model due to the new data being stored.
FormData stores a DecisionMap and a FilledFormFields, and FormDataStore
stores each user's FormData.

I decided to hand a Decision to the backdrop, instead of a Node. The
Decision comes with a Node, so we still get the Node, but it also
potentially comes with rawInput, which represents any existing data
stored in the decision. If the user revisits a previous decision,
they'll be able to see what their inputs were.