Implemented new Node types, new paths in form
Added a couple of paths into the Change Order form, which required the
addition of a new node type (FeeNode) and the modification of another
(FieldsNode).

FeeNode is a node that acts like a note and has a pre-filled numerical
field.

Added a FilledField within FieldsNode, for nodes that have both
pre-filled (not editable) and fillable fields.