Implement more of the Change Order form
Create FieldTableNode to represent a table of fields, specifically for
"Items to be Returned" in the Change Order form.

In the process of creating FieldTableNode, wanted to use static nested
classes from FieldsNode, so moved them into their own separate files,
and put all that stuff in the tree.fields package.

Implemented pretty much the rest of the Change Order Form, but still
have to work out how to solve a few issues, and then test and debug,
and fill in descriptions for the rest of the nodes.