Refactor field names
Names in the pdf are now chosen so that they will be good database
column names as well as good field names. That way, I don't need to
have a name for the pdf field, and another name for the database column
that corresponds to it. I can simply write one name, and it works in
both the pdf and in the database. Less repetition. Yay.

This PDF revision also includes name fields split into two parts, as
well as the "reason" field adjusted for better display.