Refactor form output methods
Controller is now responsible for creating temp files. I felt this made
sense since it was already responsible for deleting them.
CMSGuidedFormFill is now just responsible for filling the fields in the
form, not creating the file.

Better error handling included as a result. We return an internal
server  error if the file couldn't be created.

Renamed getFormOutput to getFilledForm.