Refactor Decision-making behavior to improve clarity.
Move main decision-making logic into DecisionTree, which simplifies the
code in general and makes CMSGuidedFormFill's makeDecision method much
easier to read.

Also makes more sense from a data structures perspective to have
DecisionTree be responsible for creating new decisions, so now more of
the logic is located in one place, making it easier to understand.

DecisionTree and FormData now apply to a particular form instance,
which is perfectly consistent with the way those classes are used and
serves to clarify that intent.

Rename Decision.rawInput to serializedInput. rawInput now refers to an
unprocessed map of input keys to input values.