Implement save-to-disk
Save decisions to file, save filled form fields to database.
- Re-implement FormDataStore to take form name into account
- Add saveToDisk method to FormData
- Add employeeId fields to FormData (this is what motivated changes to
authentication)
- Serialize decision tree as JSON

Generalize formFileName to just formName. This is related to the
earlier commit where I changed the name of ChangeOrderForm.pdf to
change_order.pdf. The reason is that I want the file name and the name
of the database table to be the same. By generalizing formFileName to
formName, I can then reference formName in the application to get both
the name of the file and the name of the database table.

Rename User to Employee