Fix bug that would prevent next decision from being created.
In an optimization that I did earlier, I had an issue where comparison
between old input and new input would result in a null pointer
exception.

I fixed it in this commit: e122eb183433a3019db009732a2748ec85052d30

Normally, when we make a decision, we create the next one as well so
that we can conveniently link it to the one we just made. At that
time, we only know what the next node's context and previous-link
should be. We can't determine its next-link or serializedInput, so
they remain null until the user makes a decision for that context.
Once the user provides input for that decision, then it's supposed to
get updated with a new serializedInput, and then we can determine its
next-link. However, before deciding to update it, we also do a
comparison to see if the old input is different from the new input.
Since the initial "input" is null, this would result in a null pointer
exception during the comparison.

I initially solved the issue by using the empty string as a default
value for serializedInput to ensure that a null value would not be
encountered during the comparison. This turned out to be a bad idea
because the empty string is the only input given for NoteNode and
NoteChecksABoxNode. When the user would make a decision for one of
those Nodes, the new input would be the same as the existing default
input, so we would detect no change, and therefore we would not create
the next decision. The next-link would remain null and then that would
result in a null pointer exception.

To correct this, I simply changed the default value back to null, and
then included a check for null value in the inputChanged comparison.
Now the initial value is different from the new input, so the next
decision will be created. I hate null checks, but this seems to be the
cleaner solution.