Implement Form Variables
Replace hard-coded values in a CMSForm with variable values.

View form variables in a generic template.

Currently, the only variables are fees, and a percentage value.

This was tricky to implement. I had two obvious potential solutions to
choose from.

One way would have been to directly update the fee value stored by each
fee node. Many nodes could potentially refer different instances of the
same value, so instead of updating each node individually, we would
probably want to have them refer to a variable value that updates
automatically. This approach would probably be the most sensible and
correct, but I think it would result in some potentially wild changes to
fundamental code. Do I now support form variables for all kinds of
nodes? It could cost me a lot to make that change.

The way I chose to do it instead, is just to completely trash all of
the nodes and recreate the entire form object from scratch. All
variable values are updated from the database when the change order
form is constructed.

I updated the singleton pattern to have a reset method that can throw
out the old instance, update values, and replace it with a new one.