Fix bug in updating form variables
CMSFormFactory#updateFormVariables was not performing a reset on
ChangeOrderForm, so ChangeOrderForm#updateFormVariables would update
the static values in ChangeOrderForm.Variable, but the instance data
wouldn't change.

Furthermore, a number of static expressions were dependent upon
Expr.adminFeePercentage. Updating the value of adminFeePercentage had
no effect on the values of the expressions that depended upon it, so
when we used those expressions to create new nodes, those nodes would
still be using the old value of adminFeePercentage.

The fix for this issue was to simply refresh those expressions during a
reset.